Separate file for search component
